Added StableHungarianMatcher
